+++ src/showinfm/system/__init__.py
@@ -7,14 +7,18 @@ from . import linux
 
 system = platform.system()
 is_wsl = False
+is_wsl1 = False
+is_wsl2 = False
 if system == "Windows":
     current_platform = Platform.windows
 elif system == "Linux":
+    current_platform = Platform.linux
     if linux.detect_wsl():
-        current_platform = Platform.windows
         is_wsl = True
-    else:
-        current_platform = Platform.linux
+        if linux.wsl_version() == linux.LinuxDesktop.wsl2:
+            is_wsl2 = True
+        else:
+            is_wsl1 = True
 elif system == "Darwin":
     current_platform = Platform.macos
 else: